requirements.txt
streamlit
pandas
numpy
sklearn
openai
plotly.express